+++ Cargo.toml
@@ -23,7 +23,6 @@ members = [
     "crates/iceberg",
     "crates/integration_tests",
     "crates/integrations/*",
-    "crates/puffin",
     "crates/test_utils",
 ]
 exclude = ["bindings/python"]
