+++ crates/iceberg/src/lib.rs
@@ -84,3 +84,5 @@ mod runtime;
 pub mod arrow;
 mod utils;
 pub mod writer;
+
+pub mod puffin;
