+++ crates/iceberg/src/puffin/blob.rs
