+++ crates/iceberg/testdata/puffin/java-generated/empty-puffin-uncompressed.bin
