+++ crates/iceberg/testdata/puffin/java-generated/sample-metric-data-compressed-zstd.bin
