+++ crates/iceberg/testdata/puffin/java-generated/sample-metric-data-uncompressed.bin
