+++ crates/iceberg/testdata/puffin/rust-generated/empty-puffin-uncompressed.bin
