+++ crates/iceberg/testdata/puffin/rust-generated/sample-metric-data-compressed-zstd.bin
