+++ crates/iceberg/testdata/puffin/rust-generated/sample-metric-data-uncompressed.bin
