--- crates/puffin/Cargo.toml
@@ -1,41 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-[package]
-name = "iceberg-puffin"
-version = { workspace = true }
-edition = { workspace = true }
-homepage = { workspace = true }
-rust-version = { workspace = true }
-
-categories = ["database"]
-description = "Apache Iceberg Puffin"
-repository = { workspace = true }
-license = { workspace = true }
-keywords = ["iceberg", "puffin"]
-
-[dependencies]
-bytes = { workspace = true }
-iceberg = { workspace = true }
-once_cell = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-zstd = { workspace = true }
-
-[dev-dependencies]
-tempfile = { workspace = true }
-tokio = { workspace = true }
